MAINT: remove duplicated conftest
