+++ conftest.py
@@ -1,8 +1,11 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-import numpy as np
-from astropy.utils import minversion
+import os
+from pathlib import Path
 
+from astropy.utils import minversion
+import numpy as np
+import pytest
 from pytest_astropy_header.display import (PYTEST_HEADER_MODULES,
                                            TESTED_VERSIONS)
 
@@ -22,6 +25,10 @@ def pytest_configure(config):
     PYTEST_HEADER_MODULES['astropy-healpix'] = 'astropy_healpix'
     PYTEST_HEADER_MODULES['vamdclib'] = 'vamdclib'
 
+    del PYTEST_HEADER_MODULES['h5py']
+    del PYTEST_HEADER_MODULES['Scipy']
+    del PYTEST_HEADER_MODULES['Pandas']
+
     # keyring doesn't provide __version__ any more
     # PYTEST_HEADER_MODULES['keyring'] = 'keyring'
 
@@ -32,3 +39,24 @@ def pytest_configure(config):
 
     TESTED_VERSIONS['astroquery'] = version.version
     TESTED_VERSIONS['astropy_helpers'] = version.astropy_helpers_version
+
+
+def pytest_addoption(parser):
+    parser.addoption(
+        '--alma-site',
+        action='store',
+        default='almascience.eso.org',
+        help='ALMA site (almascience.nrao.edu, almascience.eso.org or '
+             'almascience.nao.ac.jp for example)'
+    )
+
+
+@pytest.fixture(scope='function')
+def tmp_cwd(tmp_path):
+    """Perform test in a pristine temporary working directory."""
+    old_dir = Path.cwd()
+    os.chdir(tmp_path)
+    try:
+        yield tmp_path
+    finally:
+        os.chdir(old_dir)
